Update delete, throw error if resource not found

See merge request cpsc3192017w2/methanex/CH3OH/management-solution!36

+++ server/src/main/java/com/ch3oh/portfolio/service/SkillCategoryServiceImpl.java
@@ -67,6 +67,10 @@ public class SkillCategoryServiceImpl {
 
     @Transactional
     public void deleteSkillCategory(String id) {
+        if (!skillCategoryDao.exists(Integer.valueOf(id))) {
+            throw new GeneralRestNotFoundException();
+        }
+
         skillCategoryDao.delete(Integer.valueOf(id));
     }
